# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phoenix Orders API",
        "description": "Documentación de la API para el MVP de Phoenix Orders",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/customers": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar Clientes Activos",
                "description": "Retorna una lista de los clientes que se encuentran activos.",
                "operationId": "getCustomersList",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear Cliente",
                "description": "Registra un nuevo cliente en el sistema.",
                "operationId": "storeCustomer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Nicole Fernandez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nicole@phoenix.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+59177777777"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado"
                    },
                    "422": {
                        "description": "Error de validación de datos"
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Consultar Cliente",
                "description": "Devuelve la información de un cliente utilizando su UUID.",
                "operationId": "getCustomerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del Cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar Pedidos",
                "description": "Retorna una lista de todos los pedidos registrados en el sistema.",
                "operationId": "getOrdersList",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear Pedido",
                "description": "Crea un nuevo pedido con sus respectivos productos utilizando el OrderService transaccional.",
                "operationId": "storeOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "items"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "description": "UUID del Cliente activo",
                                        "type": "string",
                                        "example": "9a8b7c6d-5e4f-3a2b-1c0d-9e8f7a6b5c4d"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "string",
                                                    "example": "1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación (Ej. Stock insuficiente, Cliente inactivo)"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Consultar Pedido",
                "description": "Devuelve los detalles de un pedido específico y sus ítems.",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del Pedido",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar Productos Activos",
                "description": "Retorna una lista de los productos disponibles y activos.",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear Producto",
                "description": "Registra un nuevo producto en el inventario.",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Teclado Mecánico"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Teclado RGB switch blue"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "422": {
                        "description": "Error de validación de datos"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Consultar Producto",
                "description": "Devuelve la información de un producto utilizando su UUID.",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del Producto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Clientes",
            "description": "Clientes"
        },
        {
            "name": "Pedidos",
            "description": "Pedidos"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}